feat: Refactoring the homepage code
