+++ package.json
@@ -19,7 +19,6 @@
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.25.1",
     "react-icons": "^4.3.1",
-    "react-intersection-observer": "^8.33.1",
     "react-router-dom": "5",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.0",
