+++ src/routes/index.jsx
@@ -9,7 +9,6 @@ import { Favorites } from "../pages/Favorites";
 import { Phrases } from "../pages/Phrases";
 import { Comments } from "../pages/Comments";
 import { MyComments } from "../pages/MyComments";
-import { Creators } from "../pages/Creators";
 
 export const Routes = () => (
   <Switch>
@@ -21,6 +20,5 @@ export const Routes = () => (
     <Route path="/favorites" component={Favorites} isPrivate />
     <Route path="/comments/:id" component={Comments} isPrivate />
     <Route path="/mycomments" component={MyComments} isPrivate />
-    <Route path="/creators" component={Creators} />
   </Switch>
 );
