--- src/utils/creators.js
@@ -1,16 +1,0 @@
-import guilhermeAvatar from "../../assets/guilherme.svg";
-import guilhermePic from "../../assets/guilherme-pic.jpg";
-
-export const creators = [
-  {
-    id: 1,
-    name: "Guilherme Couto",
-    role: "Tech Leader",
-    avatar: guilhermeAvatar,
-    image: guilhermePic,
-    github: "https://github.com/GuiCoutoSt",
-    linkedin: "https://www.linkedin.com/in/guilhermecoutodev",
-    description:
-      "Tenho 26 anos, gosto de café expresso, dias ensolarados e deprojetos que fazem a diferença.",
-  },
-];
